+++ processor/src/test/resources/fixtures/org/mapstruct/ap/test/source/constants/ConstantsMapperImpl.java
@@ -37,8 +37,8 @@ public class ConstantsMapperImpl implements ConstantsMapper {
         constantsTarget.setFloatValue( 1.40e-45f );
         constantsTarget.setFloatBoxed( 3.4028235e38f );
         constantsTarget.setDoubleValue( 1e137 );
-        constantsTarget.setDoubleBoxed( (double) 0x0.001P-1062d );
-        constantsTarget.setDoubleBoxedZero( (double) 0.0 );
+        constantsTarget.setDoubleBoxed( 0x0.001P-1062d );
+        constantsTarget.setDoubleBoxedZero( 0.0 );
 
         return constantsTarget;
     }
